[#2187] Add usage scale value type
Adds a usage scale value type that tracks number of uses and the
recovery period. This can be used to set the uses of an item and
will display better in the class journal page.

Additional work needs to be done to allow for changing the recvoery
period based on the scale value.