[#5647] Advance bastion orders based on calendar
Automatically advance bastion orders based on when new days occur
using the built-in calendar. Allows orders to complete without
needing a bastion turn to be triggered. Each facility with an
order in progress stores the time it was last updated and that is
used to advance or complete the order when time passes.

 ### Todo
- [ ] Prompt for bastion turns on regular interval